// src/app/enum/acrecimo.atributo.enum.ts
export enum AcrecimoAtributo {
    FORCA = 'this.personagem.atributos.forca_racial = this.personagem.atributos.forca_racial + 1',
    DESTREZA = 'this.personagem.atributos.destreza_racial = this.personagem.atributos.destreza_racial + 1',
    CONSTITUICAO = 'this.personagem.atributos.constituicao_racial = this.personagem.atributos.constituicao_racial + 1',
    INTELIGENCIA = 'this.personagem.atributos.inteligencia_racial = this.personagem.atributos.inteligencia_racial + 1',
    SABEDORIA = 'this.personagem.atributos.sabedoria_racial = this.personagem.atributos.sabedoria_racial + 1',
    CARISMA = 'this.personagem.atributos.carisma_racial = this.personagem.atributos.carisma_racial + 1',
  }